refactor: split trigger tests.rs into focused matching, serde, types, and lifecycle modules [OPE-480]
