+++ crates/opengoose-teams/src/triggers/tests/mod.rs
@@ -1,0 +1,4 @@
+mod lifecycle;
+mod matching;
+mod serde_roundtrips;
+mod types;
